refactor display label validator
