+++ backend/infrahub/core/schema/schema_branch.py
@@ -63,6 +63,7 @@ from infrahub.core.schema.attribute_parameters import (
 from infrahub.core.schema.attribute_schema import get_attribute_schema_class_for_kind
 from infrahub.core.schema.definitions.core import core_profile_schema_definition
 from infrahub.core.validators import CONSTRAINT_VALIDATOR_MAP
+from infrahub.core.validators.schema_branch.display_label_validator import DisplayLabelValidator
 from infrahub.core.validators.schema_branch.hierarchical_nodes_restricted_words_validator import (
     HierarchicalNodesRestrictedWords,
 )
@@ -123,7 +124,7 @@ class SchemaBranch:
             self.profiles = data.get("profiles", {})
             self.templates = data.get("templates", {})
 
-        self.validators: list[SchemaBranchValidator] = [HierarchicalNodesRestrictedWords()]
+        self.validators: list[SchemaBranchValidator] = [HierarchicalNodesRestrictedWords(), DisplayLabelValidator()]
 
     @classmethod
     def validate(cls, data: Any) -> Self:
@@ -661,8 +662,6 @@ class SchemaBranch:
         self.validate_identifiers()
         self.sync_uniqueness_constraints_and_unique_attributes()
         self.validate_uniqueness_constraints()
-        self.validate_display_labels()
-        self.validate_display_label()
         self.validate_order_by()
         self.validate_default_filters()
         self.validate_parent_component()
@@ -886,78 +885,6 @@ class SchemaBranch:
                         | SchemaElementPathType.REL_ONE_MANDATORY_NO_ATTR,
                         element_name=element_name,
                     )
-
-    def validate_display_label(self) -> None:
-        self.display_labels = DisplayLabels()
-        inherited_display_label_nodes: set[str] = set()
-        for name in self.all_names:
-            node_schema = self.get(name=name, duplicate=False)
-
-            if node_schema.display_label is None and node_schema.display_labels:
-                update_candidate = self.get(name=name, duplicate=True)
-                if len(node_schema.display_labels) == 1:
-                    # If the previous display_labels consist of a single attribute convert
-                    # it to an attribute based display label
-                    update_candidate.display_label = _format_display_label_component(
-                        component=node_schema.display_labels[0]
-                    )
-                else:
-                    # If the previous display label consists of multiple attributes
-                    # convert it to a Jinja2 based display label
-                    update_candidate.display_label = " ".join(
-                        [
-                            f"{{{{ {_format_display_label_component(component=display_label)} }}}}"
-                            for display_label in node_schema.display_labels
-                        ]
-                    )
-                self.set(name=name, schema=update_candidate)
-            elif (
-                node_schema.display_label is None
-                and not node_schema.display_labels
-                and isinstance(node_schema, NodeSchema)
-            ):
-                # Inherit display_label from a parent generic if exactly one defines it.
-                # Validation is skipped for inherited nodes because the generic's display_label
-                # is already validated on the generic itself.
-                inherited = [
-                    self.get(name=generic, duplicate=False).display_label
-                    for generic in node_schema.inherit_from
-                    if self.get(name=generic, duplicate=False).display_label
-                ]
-                if len(inherited) == 1:
-                    update_candidate = self.get(name=name, duplicate=True)
-                    update_candidate.display_label = inherited[0]
-                    self.set(name=name, schema=update_candidate)
-                    inherited_display_label_nodes.add(name)
-
-            node_schema = self.get(name=name, duplicate=False)
-            if not node_schema.display_label or name in inherited_display_label_nodes:
-                continue
-
-            self._validate_display_label(node=node_schema)
-
-    def validate_display_labels(self) -> None:
-        for name in self.all_names:
-            node_schema = self.get(name=name, duplicate=False)
-
-            if node_schema.display_labels:
-                for path in node_schema.display_labels:
-                    self.validate_schema_path(
-                        node_schema=node_schema,
-                        path=path,
-                        allowed_path_types=SchemaElementPathType.ATTR,
-                        element_name="display_labels",
-                    )
-            elif isinstance(node_schema, NodeSchema):
-                generic_display_labels = []
-                for generic in node_schema.inherit_from:
-                    generic_schema = self.get(name=generic, duplicate=False)
-                    if generic_schema.display_labels:
-                        generic_display_labels.append(generic_schema.display_labels)
-
-                if len(generic_display_labels) == 1:
-                    # Only assign node display labels if a single generic has them defined
-                    node_schema.display_labels = generic_display_labels[0]
 
     def validate_order_by(self) -> None:
         for name in self.all_names:
@@ -1339,53 +1266,6 @@ class SchemaBranch:
                                 f" from multiple generics {sorted([duplicate, generic_schema.kind])}"
                             )
                         defined_from_generic[attribute_key] = generic_schema.kind
-
-    def _validate_display_label(self, node: MainSchemaTypes) -> None:
-        if not node.display_label:
-            return
-
-        if not any(c in node.display_label for c in "{}"):
-            schema_path = self.validate_schema_path(
-                node_schema=node,
-                path=node.display_label,
-                allowed_path_types=SchemaElementPathType.ATTR_WITH_PROP,
-                element_name="display_label - non Jinja2",
-            )
-            if schema_path.attribute_schema and node.is_node_schema and node.namespace not in ["Internal", "Schema"]:
-                self.display_labels.register_attribute_based_display_label(
-                    kind=node.kind, attribute_name=schema_path.attribute_schema.name
-                )
-            return
-
-        jinja_template = InfrahubJinja2Template(template=node.display_label)
-        context = ExecutionContext.CORE
-        if not config.SETTINGS.security.restrict_untrusted_jinja2_filters:
-            context |= ExecutionContext.LOCAL
-        try:
-            variables = jinja_template.get_variables()
-            jinja_template.validate(context=context)
-        except (JinjaTemplateOperationViolationError, JinjaTemplateError) as exc:
-            raise ValueError(
-                f"{node.kind}: display_label is set to a jinja2 template, but has an invalid template: {exc.message}"
-            ) from exc
-
-        allowed_path_types = (
-            SchemaElementPathType.ATTR_WITH_PROP
-            | SchemaElementPathType.REL_ONE_MANDATORY_ATTR_WITH_PROP
-            | SchemaElementPathType.REL_ONE_ATTR_WITH_PROP
-        )
-        for variable in variables:
-            schema_path = self.validate_schema_path(
-                node_schema=node, path=variable, allowed_path_types=allowed_path_types, element_name="display_label"
-            )
-
-            if schema_path.is_type_attribute and schema_path.active_attribute_schema.name == "display_label":
-                raise ValueError(f"{node.kind}: display_label the '{variable}' variable is a reference to itself")
-
-            if node.is_node_schema and node.namespace not in ["Internal", "Schema"]:
-                self.display_labels.register_template_schema_path(
-                    kind=node.kind, schema_path=schema_path, template=node.display_label
-                )
 
     def _validate_computed_attribute(self, node: NodeSchema, attribute: AttributeSchema) -> None:
         if not attribute.computed_attribute or attribute.computed_attribute.kind == ComputedAttributeKind.USER:
@@ -2901,16 +2781,3 @@ class SchemaBranch:
                 updated_used_by_node = set(chain(template_schema_kinds, set(core_node_schema.used_by)))
                 core_node_schema.used_by = sorted(updated_used_by_node)
                 self.set(name=InfrahubKind.NODE, schema=core_node_schema)
-
-
-def _format_display_label_component(component: str) -> str:
-    """Return correct format for display_label.
-
-    Previously both the format of 'name' and 'name__value' was
-    supported this function ensures that the proper 'name__value'
-    format is used
-    """
-    if "__" in component:
-        return component
-
-    return f"{component}__value"
